src/main/java/com/mongodbdemo/kitchensink/service/SequenceGeneratorService.java
package com.mongodbdemo.kitchensink.service;

import com.mongodbdemo.kitchensink.model.DatabaseSequence;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.data.mongodb.core.FindAndModifyOptions;
import org.springframework.data.mongodb.core.MongoOperations;
import org.springframework.data.mongodb.core.query.Criteria;
import org.springframework.data.mongodb.core.query.Query;
import org.springframework.data.mongodb.core.query.Update;
import org.springframework.stereotype.Service;

import java.util.Optional;

/**
 * Service for generating unique sequences in MongoDB.
 */
@Service
public class SequenceGeneratorService {

    private final MongoOperations mongoOperations;

    @Autowired
    public SequenceGeneratorService(MongoOperations mongoOperations) {
        this.mongoOperations = mongoOperations;
    }

    /**
     * Generates a new sequence number for the given sequence name.
     *
     * @param seqName the name of the sequence
     * @return the generated sequence number
     */
    public long generateSequence(String seqName) {
        Query query = new Query(Criteria.where("_id").is(seqName));
        Update update = new Update().inc("seq", 1);

        DatabaseSequence counter = mongoOperations.findAndModify(
                query,
                update,
                FindAndModifyOptions.options().returnNew(true).upsert(true),
                DatabaseSequence.class
        );

        return Optional.ofNullable(counter)
                .map(DatabaseSequence::getSeq)
                .orElse(1L);
    }
}
